+++ packages/map-template/src/hooks/useDirectionsInfo.js
@@ -14,11 +14,11 @@ const useDirectionsInfo = (originLocation, destinationLocation, directionsServic
     const [totalTime, setTotalTime] = useState();
     const [hasFoundRoute, setHasFoundRoute] = useRecoilState(hasFoundRouteState);
     const [, setDirectionsResponse] = useRecoilState(directionsResponseState);
-    const [isDirectionReady, setIsDirectionReady] = useState();
     const accessibility = useRecoilValue(accessibilityState);
+    const [areDirectionsReady, setAreDirectionReady] = useState();
 
     useEffect(() => {
-        setIsDirectionReady(false);
+        setAreDirectionReady(false);
         if (originLocation?.geometry && destinationLocation?.geometry) {
             directionsService.getRoute({
                 origin: getLocationPoint(originLocation),
@@ -27,7 +27,6 @@ const useDirectionsInfo = (originLocation, destinationLocation, directionsServic
                 avoidStairs: accessibilityOn || accessibility
             }).then(directionsResult => {
                 if (directionsResult && directionsResult.legs) {
-                    setHasFoundRoute(true);
                     // Calculate total distance and time
                     const totalDistance = directionsResult.legs.reduce((accumulator, current) => accumulator + current.distance.value, 0);
                     const totalTime = directionsResult.legs.reduce((accumulator, current) => accumulator + current.duration.value, 0);
@@ -42,7 +41,7 @@ const useDirectionsInfo = (originLocation, destinationLocation, directionsServic
                         totalTime,
                         directionsResult
                     });
-                    setIsDirectionReady(true);
+                    setAreDirectionReady(true);
                 } else {
                     setHasFoundRoute(false);
                 }
@@ -52,7 +51,7 @@ const useDirectionsInfo = (originLocation, destinationLocation, directionsServic
         }
     }, [originLocation, destinationLocation, directionsService, accessibilityOn, travelMode]);
 
-    return [totalDistance, totalTime, hasFoundRoute, isDirectionReady];
+    return [totalDistance, totalTime, hasFoundRoute, areDirectionsReady];
 }
 
 export default useDirectionsInfo;