+++ packages/map-template/CHANGELOG.md
@@ -5,12 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.53.0] - 2024-06-19
+## [1.54.0] - 2024-07-02
 
 ### Added
 
 - Added a new URL parameter `skipGo`: A boolean parameter that dictates if directions 'Go' button should be skipped, showing the route straightaway.
 - Added a new URL parameter `accessibility`: A boolean parameter that dictates if accessibility should be taken into consideration when getting a route.
+
+## [1.53.1] - 2024-07-02
+
+### Fixed
+
+- Wayfinding 'Go' button is now disabled until route is fetched and ready to be displayed.
+
+## [1.53.0] - 2024-07-01
+
+### Added
+
+- Add support for new parameter `showExternalIDs` which determines if the Location details should show the External ID or not.
 
 ## [1.52.0] - 2024-06-19
 
@@ -52,7 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed an issue with clearing search input causing console errrors.
+- Fixed an issue with clearing search input causing console errors.
 
 ## [1.48.0] - 2024-06-13